% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link.R
\name{add_link}
\alias{add_link}
\title{Add other association link plot on correlation plot.}
\usage{
add_link(df, mapping = NULL, spec.key = "spec", env.key = "env",
  curvature = NULL, spec.label.hspace = NULL, spec.label.vspace = 0,
  on.left = FALSE, diag.label = FALSE, extra.params = extra_params(),
  ...)
}
\arguments{
\item{df}{a data frame object.}

\item{mapping}{NULL (default) or a list of aesthetic mappings to use for plot.}

\item{spec.key}{string (defaults to "spec"), group variables names in \code{df}.}

\item{env.key}{string (defaults to "env"),  variables names in \code{df} that
associated with the correlation coefficient matrix.}

\item{curvature}{a numeric value giving the amount of curvature.}

\item{spec.label.hspace, spec.label.vspace}{a numeric value giving the amount of
horizontal/vertical space betweed group points and labels.}

\item{on.left}{add link plot on left or right when the type of correlation plot
is "full".}

\item{diag.label}{logical (defaults to FALSE) to indicate whether add diag labels.}

\item{extra.params}{other parameters that control link details can only be set using
the \code{\link[ggcor]{extra_params}} function.}

\item{...}{extra params passing to \code{\link[ggplot2]{geom_curve}}.}
}
\description{
This function can add other associated information link plot more
    quickly, and this function can be used to reflect the relationship between
    other variables and variables in the correlation coefficient matrix plot.
}
\examples{
library(vegan)
data("varechem")
data("varespec")
mantel <- fortify_mantel(varespec, varechem,
                           spec.select = list(1:10, 5:14, 7:22, 9:32))
quickcor(varechem, type = "upper") +
  geom_square() +
  add_link(mantel, diag.label = TRUE) +
  add_diag_label() + remove_axis("x")
library(dplyr)
mantel01 <- mantel \%>\%
  mutate(r = cut(r, breaks = c(-Inf, 0.25, 0.5, Inf),
                 labels = c("<0.25", "0.25-0.5", ">=0.5"),
                 right = FALSE),
         p.value = cut(p.value, breaks = c(-Inf, 0.001, 0.01, 0.05, Inf),
                       labels = c("<0.001", "0.001-0.01", "0.01-0.05", ">=0.05"),
                       right = FALSE))
quickcor(varechem, type = "upper") + geom_square() +
  add_link(mantel01, mapping = aes(colour = p.value, size = r),
           diag.label = TRUE) +
  add_diag_label() +
  scale_size_manual(values = c(0.5, 1.5, 3)) +
  remove_axis("x")
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
