% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_bar}
\alias{anno_bar}
\alias{anno_bar2}
\alias{anno_boxplot}
\title{Bar and boxplot annotation}
\usage{
anno_bar(
  data,
  mapping,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)

anno_bar2(
  data,
  mapping,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)

anno_boxplot(
  data,
  mapping,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{align}{align base on main plot.}

\item{flip}{a logical value. If TRUE, the annotation on the left or on
the bottom side will be flipped.}

\item{pos}{position of annotation.}

\item{width, height}{width and height of annotation.}

\item{trans}{the name of a transformation object or the object itself.}

\item{scale}{scale object.}

\item{theme}{theme object or the name.}

\item{...}{extra parameters.}
}
\value{
a anno_* object.
}
\description{
Bar and boxplot annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
