% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_row_heat}
\alias{anno_row_heat}
\alias{anno_col_heat}
\title{Heatmap annotation}
\usage{
anno_row_heat(
  data,
  mapping = NULL,
  align = TRUE,
  geom = "tile",
  mark = NULL,
  space = 0.5,
  width = 0.2,
  label = TRUE,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  ...
)

anno_col_heat(
  data,
  mapping = NULL,
  align = TRUE,
  geom = "tile",
  mark = NULL,
  space = 0.5,
  height = 0.2,
  label = TRUE,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{align}{align base on main plot.
the bottom side will be flipped.}

\item{geom}{one of "anno_tile", "anno_tile2" or "point".}

\item{mark}{a layer instance.}

\item{space}{scala numeric value.}

\item{width, height}{width and height of annotation.}

\item{label}{logical value, if TRUE will add label on plot.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{...}{extra parameters.}
}
\value{
a anno_*_heat object.
}
\description{
Heatmap annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
