% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R, R/deprecated.R
\name{anno_link}
\alias{anno_link}
\alias{add_link}
\alias{anno_link_label}
\title{Link annotation}
\usage{
anno_link(
  data,
  mapping = NULL,
  width = 0.3,
  pos = "right",
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  nudge_x = 0.1,
  expand = NULL,
  start.var = NULL,
  start.name = NULL,
  end.var = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{width}{width of link plot.}

\item{pos}{position of link plot.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{nudge_x}{horizonal justification of label.}

\item{expand}{expansion of x axis.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
will be converted.}

\item{start.name}{character vector.}

\item{...}{extra parameters passing to layer function.}
}
\value{
a ggplot layer.
}
\description{
Link annotation
}
\note{
Loading the \code{ggraph} package first makes this function even more functional.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
