% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_bar}
\alias{anno_bar}
\alias{anno_bar2}
\alias{anno_boxplot}
\alias{anno_row_heat}
\alias{anno_col_heat}
\title{Custom annotation}
\usage{
anno_bar(
  data,
  mapping,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)

anno_bar2(
  data,
  mapping,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)

anno_boxplot(
  data,
  mapping,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  scale = NULL,
  theme = NULL,
  ...
)

anno_row_heat(
  data,
  mapping = NULL,
  align = TRUE,
  geom = "anno_tile",
  mark = NULL,
  space = 0.5,
  width = 1,
  label = TRUE,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  ...
)

anno_col_heat(
  data,
  mapping = NULL,
  align = TRUE,
  geom = "anno_tile2",
  mark = NULL,
  space = 0.5,
  height = 1,
  label = TRUE,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{align}{align base on main plot.}

\item{flip}{a logical value. If TRUE, the annotation on the left or on
the bottom side will be flipped.}

\item{pos}{position of annotation.}

\item{width, height}{width and height of annotation.}

\item{trans}{the name of a transformation object or the object itself.}

\item{scale}{scale object.}

\item{theme}{theme object or the name.}

\item{...}{extra parameters.}

\item{geom}{one of "anno_tile", "anno_tile2" or "point".}

\item{mark}{a layer instance.}

\item{space}{scala numeric value.}

\item{label}{logical value, if TRUE will add label on plot.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}
}
\value{
anno_* object.
}
\description{
Custom annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
