% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R, R/deprecated.R
\name{anno_row_tree}
\alias{anno_row_tree}
\alias{anno_col_tree}
\alias{anno_tree}
\title{Annotation for correlation matrix plot}
\usage{
anno_row_tree(
  bcols = NULL,
  width = NULL,
  pos = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color
)

anno_col_tree(
  bcols = NULL,
  height = NULL,
  pos = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color
)
}
\arguments{
\item{bcols}{branch colours.}

\item{width, height}{width or height of tree.}

\item{pos}{position of tree.}

\item{colour, color}{colour of segments.}

\item{size}{width of segments.}

\item{linetype}{line type of segments.}
}
\value{
anno_tree object.
}
\description{
Annotation for correlation matrix plot
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
