% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cor-tbl.R
\name{as_cor_tbl}
\alias{as_cor_tbl}
\alias{as_cor_tbl.matrix}
\alias{as_cor_tbl.data.frame}
\alias{as_cor_tbl.correlate}
\alias{as_cor_tbl.rcorr}
\alias{as_cor_tbl.corr.test}
\alias{as_cor_tbl.mantel_tbl}
\alias{as_cor_tbl.pro_tbl}
\alias{as_cor_tbl.default}
\title{Coerce to a cor_tbl object}
\usage{
as_cor_tbl(x, ...)

\method{as_cor_tbl}{matrix}(x, ...)

\method{as_cor_tbl}{data.frame}(x, ...)

\method{as_cor_tbl}{correlate}(x, extra.mat = list(), ...)

\method{as_cor_tbl}{rcorr}(x, ...)

\method{as_cor_tbl}{corr.test}(x, ...)

\method{as_cor_tbl}{mantel_tbl}(x, byrow = TRUE, ...)

\method{as_cor_tbl}{pro_tbl}(x, byrow = TRUE, ...)

\method{as_cor_tbl}{default}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{extra params passing to \code{\link[ggcor]{cor_tbl}}.}

\item{extra.mat}{any other matrix-like data with same dimmsion as \code{x}.}

\item{byrow}{a logical value indicating whether arrange the 'spec' columns on y axis.}
}
\value{
a cor_tbl object.
}
\description{
Functions to coerce a object to cor_tbl if possible.
}
\examples{
cor(mtcars) \%>\% as_cor_tbl()
correlate(mtcars, cor.test = TRUE) \%>\% as_cor_tbl()
correlate(mtcars, type = "upper") \%>\% as_cor_tbl()
\dontrun{
## S3 method for rcorr object
require(Hmisc)
rcorr(mtcars) \%>\% as_cor_tbl()

## S3 method for corr.test object
require(psych)
corr.test(mtcars) \%>\% as_cor_tbl()
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
