% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour-pal.R
\name{brown_blue}
\alias{brown_blue}
\alias{pink_green}
\alias{purple_green}
\alias{brown_purple}
\alias{red_blue}
\alias{red_grey}
\alias{red_yellow_blue}
\alias{red_yellow_green}
\alias{spectral}
\alias{link_colour_pal}
\title{Colour pal}
\usage{
brown_blue(n = 11)

pink_green(n = 11)

purple_green(n = 11)

brown_purple(n = 11)

red_blue(n = 11)

red_grey(n = 11)

red_yellow_blue(n = 11)

red_yellow_green(n = 11)

spectral(n = 11)

link_colour_pal(n)
}
\arguments{
\item{n}{number of different colors in the palette, minimum 3, maximum 11.}
}
\value{
a palette.
}
\description{
Wrapper for the diverging palettes provided by \code{\link{RColorBrewer}}.
}
