% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-network.R
\name{cor_network}
\alias{cor_network}
\alias{print.cor_network}
\title{Tidy co-occurrence network data}
\usage{
cor_network(
  corr,
  p.value = NULL,
  directed = FALSE,
  row.names = NULL,
  col.names = NULL,
  rm.dup = TRUE,
  simplify = TRUE,
  weight = NULL,
  r.thres = 0.6,
  r.abs = TRUE,
  p.thres = 0.05,
  val.type = "tbl_graph"
)

\method{print}{cor_network}(x, n = 3, ...)
}
\arguments{
\item{corr}{correlation matrix.}

\item{p.value}{significant matrix of correlation.}

\item{directed}{logical value, whether or not to create a directed graph.}

\item{row.names, col.names}{row and column names of correlation matrix.}

\item{rm.dup}{logical (defaults to TRUE) indicating whether remove duplicate
rows. If TRUE, the correlation between A-B and B-A is retained only A-B.}

\item{simplify}{logical value (defaults to TRUE) indicating whether to
delete nodes without edge connections.}

\item{weight}{NULL (default) or name of column in edges which will be renamed
to "weight".}

\item{r.thres}{a numeric value.}

\item{r.abs}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}

\item{val.type}{type return value:
\itemize{
    \item \code{tbl_graph}: return tbl_graph object
    \item \code{igraph}: return igraph object
    \item \code{list}: return a list of nodes and edges
 }}

\item{x}{a cor_network object.}

\item{n}{number of rows to show.}

\item{...}{extra params for printing.}
}
\value{
a tbl_graph (default), igraph or list of nodes and edges.
}
\description{
The function calculates correlation coefficient, statistical
    significance level and filters according to conditions.
}
\examples{
cor_network(cor(mtcars))
corr <- correlate(mtcars, cor.test = TRUE)
cor_network(corr$r, corr$p.value)

## return a igraph object
cor_network(corr$r, corr$p.value, val.type = "igraph")

## reurn a tbl_graph object
cor_network(corr$r, corr$p.value, val.type = "tbl_graph")

}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
