% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-tbl-utils.R
\name{cor_tbl2tbl_df}
\alias{cor_tbl2tbl_df}
\title{Coerce cor_tbl to tbl_df}
\usage{
cor_tbl2tbl_df(x, rm.extra.attr = TRUE)
}
\arguments{
\item{x}{cor_tbl object}

\item{rm.extra.attr}{a logical value indicating whether remove the extra attributes.}
}
\value{
a tbl_df object.
}
\description{
Coerce cor_tbl to tbl_df
}
\examples{
df <- fortify_cor(mtcars)
cor_tbl2tbl_df(df)
}
