% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-tbl-utils.R
\name{cor_tbl_namebind}
\alias{cor_tbl_namebind}
\title{Convert cor_tbl from order to name}
\usage{
cor_tbl_namebind(x, add = FALSE, var_xname = NULL, var_yname = NULL,
  keep.factor = TRUE)
}
\arguments{
\item{x}{A cor_tbl object.}

\item{add}{A logical value indicating whether binds x/y name to a new columns.}

\item{var_xname}{If add = TRUE, specifies the name of the new x column.}

\item{var_yname}{If add = TRUE, specifies the name of the new y column.}

\item{keep.factor}{A logical value indicating whether trans x/y name column to
character vector.}
}
\value{
A modified tbl.
}
\description{
The cor_tbl x corresponds to the column of the correlation matrix,
    and y corresponds to the row of the correlation matrix. By default, x/y is
    the order in which the rows or columns appear, and this function binds x/y
    to the columns/rows name of the correlation matrix.
}
\author{
Houyun Huang
}
