% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-test.R
\name{cor_test}
\alias{cor_test}
\title{Significance Test for Correlattion.}
\usage{
cor_test(x, y = NULL, alternative = "two.sided", method = "pearson",
  ...)
}
\arguments{
\item{x}{a matrix object.}

\item{y}{a matrix object or NULL.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided",
"greater" or "less".}

\item{method}{a character string indicating which correlation coefficient is to be used
for the test. One of "pearson", "kendall", or "spearman".}

\item{...}{extra params passing to `cor.test()`.}
}
\value{
a list with P values matrix, upper of confidence intervals matrix and lower of
    confidence intervals matrix.
}
\description{
Significance test which produces p-values and confidence intervals for
    each pair of input features.
}
\details{
The columns of 'x' will be tested for each pair when y is NULL(the default),
    otherwise each column in 'x' and each column in 'y' is tested for each pair.
}
\examples{
cor_test(mtcars)
m1 <- matrix(rnorm(100), nrow = 10)
m2 <- matrix(rnorm(60), nrow = 10)
cor_test(m1, m2)
}
\seealso{
\code{\link[stats]{cor.test}}.
}
\author{
Houyun Huang

Lei Zhou

Jian Chen

Taiyun Wei
}
