% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-network-data.R
\name{create_network}
\alias{create_network}
\alias{fast_create_network}
\title{Tidy co-occurrence network data}
\usage{
create_network(x, r.thres = 0.6, r.absolute = TRUE, p.thres = 0.05,
  ...)

fast_create_network(x, r.thres = 0.6, r.absolute = TRUE,
  p.thres = 0.05, ...)
}
\arguments{
\item{x}{any \code{R} object can be converted to cor_tbl.}

\item{r.thres}{a numeric value.}

\item{r.absolute}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}

\item{...}{passing to \code{\link[ggcor]{as_cor_tbl}}}
}
\value{
cor_tbl.
}
\description{
The function calculates correlation coefficient, statistical
    significance level and filters according to conditions.
}
