% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-cor.R
\name{display_cor}
\alias{display_cor}
\alias{display_cor.matrix}
\alias{display_cor.data.frame}
\alias{display_cor.correlate}
\alias{display_cor.rcorr}
\alias{display_cor.corr.test}
\alias{display_cor.cor_tbl}
\alias{display_cor.mantel_tbl}
\alias{display_cor.pro_tbl}
\alias{format_cor}
\title{Display correlation matrix with nice format}
\usage{
display_cor(x, ...)

\method{display_cor}{matrix}(x, ...)

\method{display_cor}{data.frame}(x, ...)

\method{display_cor}{correlate}(x, ...)

\method{display_cor}{rcorr}(x, ...)

\method{display_cor}{corr.test}(x, ...)

\method{display_cor}{cor_tbl}(
  x,
  type = "full",
  show.diag = FALSE,
  digits = 2,
  nsmall = 2,
  sig.level = c(0.05, 0.01, 0.001),
  mark = c("*", "**", "***"),
  coef = NULL,
  nice.format = TRUE,
  ...
)

\method{display_cor}{mantel_tbl}(x, byrow = TRUE, ...)

\method{display_cor}{pro_tbl}(x, byrow = TRUE, ...)

format_cor(
  corr,
  p.value = NULL,
  type = "full",
  show.diag = FALSE,
  digits = 2,
  nsmall = 2,
  sig.level = c(0.05, 0.01, 0.001),
  mark = c("*", "**", "***"),
  nice.format = TRUE
)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{extra params passing to \code{format_cor}.}

\item{type}{a string, "full" (default), "upper" or "lower", display full,
lower triangular or upper triangular matrix.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}

\item{digits}{integer indicating the number of decimal places (round) or
significant digits (signif) to be used, the default value is 2.}

\item{nsmall}{the minimum number of digits to the right of the decimal
point in formatting real/complex numbers in non-scientific formats,
the default value is 2.}

\item{sig.level}{significance level，the defaults values is [0.05, 0.01, 0.001].}

\item{mark}{significance mark，the defaults values is ["*", "**", "***"].}

\item{coef}{string to specifies which column is the coefficient when "x"
is a general_cor_tbl.}

\item{nice.format}{a logical value indicating whether the output needs to be
automatically optimized.}

\item{byrow}{a logical value indicating whether arrange the 'spec' columns on y axis.}

\item{corr}{a correlation matrix.}

\item{p.value}{NULL (default) or a matrix of p value.}
}
\value{
a data frame.
}
\description{
Functions to display correlation object.

Format an correlation matrix for printing.
}
\examples{
corr <- correlate(mtcars, cor.test = TRUE)
format_cor(corr$r, corr$p.value, type = "lower")
display_cor(corr)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
