% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-axis.R
\name{expand_axis}
\alias{expand_axis}
\title{Expand axis limits}
\usage{
expand_axis(x = NULL, y = NULL)
}
\arguments{
\item{x, y}{NULL (default) or numeric vector.}
}
\description{
Force to extend the coordinate range of the ggplot object.
}
\examples{
quickcor(mtcars) + geom_square() + expand_axis(x = -3)
quickcor(mtcars) + geom_square() + expand_axis(y = 16)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
