% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-cor-network.R
\name{export_cor_network}
\alias{export_cor_network}
\alias{export_cor_network.cor_network}
\alias{export_cor_network.cor_tbl}
\alias{export_cor_network.mantel_tbl}
\alias{export_cor_network.pro_tbl}
\alias{export_cor_network.correlate}
\alias{export_cor_network.rcorr}
\alias{export_cor_network.corr.test}
\title{Output co-network object}
\usage{
export_cor_network(x, file = "", what = "edges", sep = ",", ...)

\method{export_cor_network}{cor_network}(x, file = "", what = "edges", sep = ",", ...)

\method{export_cor_network}{cor_tbl}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)

\method{export_cor_network}{mantel_tbl}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)

\method{export_cor_network}{pro_tbl}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)

\method{export_cor_network}{correlate}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  row.names = NULL,
  col.names = NULL,
  rm.dup = TRUE,
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)

\method{export_cor_network}{rcorr}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  row.names = NULL,
  col.names = NULL,
  rm.dup = TRUE,
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)

\method{export_cor_network}{corr.test}(
  x,
  file = "",
  what = "edges",
  sep = ",",
  row.names = NULL,
  col.names = NULL,
  rm.dup = TRUE,
  simplify = TRUE,
  r.thres = 0.6,
  r.absolute = TRUE,
  p.thres = 0.05,
  ...
)
}
\arguments{
\item{x}{\code{R} object can be convert to \code{cor_network}.}

\item{file}{a character string naming a file for writing. "" indicates
output to the console.}

\item{what}{either "edges" (default) or "nodes" will be output.}

\item{sep}{the field separator string (defaults to ",").}

\item{...}{extra params passing to \code{\link[utils]{write.table}}.}

\item{simplify}{logical value (defaults to TRUE) indicating whether to
delete nodes without edge connections.}

\item{r.thres}{a numeric value.}

\item{r.absolute}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}

\item{row.names, col.names}{row and column names of correlation matrix.}

\item{rm.dup}{logical (defaults to TRUE) indicating whether remove duplicate
rows. If TRUE, the correlation between A-B and B-A is retained only A-B.}
}
\description{
Functions to output edges or nodes data of co-network
    to a file.
}
\seealso{
\code{\link[utils]{write.table}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
