% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link-extra-params.R
\name{text_params}
\alias{text_params}
\alias{point_params}
\alias{link_params}
\title{Extra params elements}
\usage{
text_params(colour = "black", size = 3.88, angle = 0, hjust = NULL,
  vjust = 0.5, alpha = NA, family = "", fontface = 1,
  color = NULL)

point_params(alpha = NA, colour = "black", fill = NA, shape = 21,
  size = 1, stroke = 0.5, color = NULL)

link_params(env.point.hjust = NULL, env.point.vjust = NULL,
  spec.point.hjust = NULL, spec.point.vjust = NULL)
}
\arguments{
\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{size}{Line/border size in mm; text size in pts.}

\item{angle}{angle (in [0, 360]).}

\item{hjust}{horizontal justification (in [0, 1]).}

\item{vjust}{vertical justification (in [0, 1]).}

\item{alpha}{alpha level in [0,1].}

\item{family}{font family.}

\item{fontface}{font face.}

\item{fill}{Fill colour.}

\item{shape}{point shape.}

\item{stroke}{point stroke.}

\item{spec.point.hjust, env.point.hjust}{numeric vector, adjust horizontal position of points.}

\item{spec.point.vjust, env.point.vjust}{numeric vector, adjust vertical position of points.}
}
\description{
In conjunction with the \code{\link[ggcor]{extra_params}}.
}
\seealso{
\code{\link[ggcor]{extra_params}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
