% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_lower_data}
\alias{get_lower_data}
\alias{get_upper_data}
\alias{get_diag_tri}
\alias{get_diag_data}
\alias{is_symmet}
\title{Helper function to extract cor_tbl.}
\usage{
get_lower_data(x, show.diag = TRUE)

get_upper_data(x, show.diag = TRUE)

get_diag_tri(x)

get_diag_data(x)

is_symmet(x)
}
\arguments{
\item{x}{a cor_tbl object.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}
}
\value{
a modified cor_tbl object.
}
\description{
These functions are used to quickly obtain the upper
    trig, lower trig, diagonal, or remove the diagonal of the correlation
    coefficient matrix.
}
\examples{
df <- fortify_cor(mtcars)
quickcor(df) + geom_colour()

## exclude upper
df \%>\% get_lower_data() \%>\%
  quickcor() + geom_colour()

## exclude lower
df \%>\% get_upper_data(show.diag = FALSE) \%>\%
  quickcor() + geom_colour()

## get the diagonal
df \%>\% get_diag_data() \%>\%
  quickcor() + geom_colour()

## exclude the diagonal
df \%>\% get_diag_tri() \%>\%
  quickcor() + geom_colour()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
