% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-otu.R
\name{filter_otu}
\alias{filter_otu}
\title{Filter OTU table by relative abundance.}
\usage{
filter_otu(x, thres = 10^-5, trans.frac = FALSE)
}
\arguments{
\item{x}{a OTU table, every otu on columns.}

\item{thres}{a numeric value, OTUs less than this value are filtered.}

\item{trans.frac}{logical value indicates whether to scale the original
OTU table to relative abundance.}
}
\value{
new OTU table.
}
\description{
Lower OTUs is filtered according to relative abundance to
   reduce the dimension of OTU table.
}
