% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-cor.R
\name{fortify_cor}
\alias{fortify_cor}
\title{Convert to cor_tbl based on input type.convert}
\usage{
fortify_cor(
  x,
  y = NULL,
  is.cor = FALSE,
  group = NULL,
  type = "full",
  show.diag = TRUE,
  cor.test = FALSE,
  cluster = FALSE,
  cluster.method = "complete",
  ...
)
}
\arguments{
\item{x}{any \code{R} object.}

\item{y}{NULL (default) or a matrix or data frame with compatible
dimensions to x.}

\item{is.cor}{logical value (default to FALSE) indicating wheater
\code{x} is a correlation matrix.}

\item{group}{NULL (default) or a vector that has the same number
of rows as x.}

\item{type}{a string, "full" (default), "upper" or "lower", display full,
lower triangular or upper triangular matrix.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}

\item{cor.test}{logical value (default to TRUE) indicating whether test
for the correlation.}

\item{cluster}{logical value (default to FALSE) indicating whether reorder
the correlation matrix by cluster.}

\item{cluster.method}{the agglomeration method to be used. This should be
(an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC).}

\item{...}{extra params passing to \code{matrix_order}.}
}
\value{
cor_tbl object.
}
\description{
The fortify_cor function is a deep encapsulation of
the \code{as_cor_tbl} function and also supports converting raw
data into cor_tbl objects by calculation.
}
\examples{
fortify_cor(mtcars)
fortify_cor(iris[-5], group = iris[[5]])
fortify_cor(mtcars, type = "lower", cluster = TRUE)
m <- cor(mtcars)
fortify_cor(m, is.cor = TRUE)
}
\seealso{
\code{\link[ggcor]{matrix_order}}, \code{\link[stats]{hclust}},
    \code{\link[ggcor]{as_cor_tbl}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
