% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/geom-diag-label.R
\name{add_diag_label}
\alias{add_diag_label}
\alias{geom_diag_label}
\title{Add diagnoal labels on correlation plot}
\usage{
geom_diag_label(..., geom = "text", remove.axis = TRUE)
}
\arguments{
\item{...}{extra parameters.}

\item{geom}{one of "text", "label" or "image".}

\item{remove.axis}{if TRUE, will remove the axis.}
}
\description{
\code{geom_diag_label} is mainly used with \code{ggcor} and
    \code{quickcor} functions to add diagnoal labels on correct position
    base on different type of cor_tbl object.
}
\examples{
quickcor(mtcars, type = "upper") + geom_colour() + geom_diag_label()
quickcor(mtcars, type = "lower") + geom_colour() + geom_diag_label()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
