% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-hc-rect.R
\name{geom_hc_rect}
\alias{geom_hc_rect}
\alias{get_hc_rect_df}
\title{Draw square mark on correlation matrix plot}
\usage{
geom_hc_rect(data = NULL, fill = NA, colour = "black", size = 2, color = NULL)

get_hc_rect_df()
}
\arguments{
\item{data}{a correlation matrix or hc_rect_df object.}

\item{fill}{NA (default) or the fill colour of square.}

\item{colour, color}{the colour of square boder.}

\item{size}{size of square boder line.}
}
\description{
Draw the cluster square mark on the correlation matrix plot.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
