% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-link.R
\name{geom_link}
\alias{geom_link}
\alias{geom_link_point}
\alias{geom_link_label}
\alias{geom_start_point}
\alias{geom_end_point}
\alias{geom_start_label}
\alias{geom_end_label}
\alias{get_start_nodes}
\alias{get_end_nodes}
\title{Special layer function for correlation link plot}
\usage{
geom_link(
  mapping = NULL,
  data = NULL,
  curvature = 0,
  inherit.aes = getOption("ggcor.link.inherit.aes", TRUE),
  ...
)

geom_link_point(...)

geom_link_label(...)

geom_start_point(
  mapping = NULL,
  data = NULL,
  inherit.aes = getOption("ggcor.link.inherit.aes", TRUE),
  ...
)

geom_end_point(
  mapping = NULL,
  data = NULL,
  inherit.aes = getOption("ggcor.link.inherit.aes", TRUE),
  ...
)

geom_start_label(
  mapping = NULL,
  data = NULL,
  inherit.aes = getOption("ggcor.link.inherit.aes", TRUE),
  ...
)

geom_end_label(
  mapping = NULL,
  data = NULL,
  inherit.aes = getOption("ggcor.link.inherit.aes", TRUE),
  ...
)

get_start_nodes()

get_end_nodes()
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{NULL or a layout_link_tbl object that create by
\code{parallel_layout()} or \code{combination_layout()}.}

\item{curvature}{a numeric value giving the amount of curvature.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{...}{extra parameters passing to layer function.}
}
\value{
geom layer.
}
\description{
A set of custom layer functions that quickly add
layers of curves, nodes, and labels.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
