% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-link.R
\name{geom_links}
\alias{geom_links}
\alias{geom_links_label}
\title{Special layer function for correlation link plot}
\usage{
geom_links(
  mapping = NULL,
  curvature = 0,
  layout = "parallel",
  layout.params = list(),
  ...
)

geom_links_label(mapping = NULL, geom = "text", ...)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{curvature}{a numeric value giving the amount of curvature.}

\item{...}{extra parameters passing to layer function.}
}
\value{
a ggplot layer.
}
\description{
A set of custom layer functions that quickly add
layers of curves, nodes, and labels.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
