% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cor-tbl.R
\name{get_row_name}
\alias{get_row_name}
\alias{get_col_name}
\alias{get_type}
\alias{get_show_diag}
\alias{nrows}
\alias{ncols}
\alias{is_cor_tbl}
\alias{is_gcor_tbl}
\title{Helper function for cor_tbl}
\usage{
get_row_name(x)

get_col_name(x)

get_type(x)

get_show_diag(x)

nrows(x)

ncols(x)

is_cor_tbl(x)

is_gcor_tbl(x)
}
\arguments{
\item{x}{a cor_tbl.}
}
\value{
return attribute value.
}
\description{
Helper function for cor_tbl
}
\examples{
df <- fortify_cor(mtcars)
## get rows names
get_row_name(df)

## get columns names
get_col_name(df)

## get show.diag parameter
get_show_diag(df)

## get type parameter
get_type(df)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
