% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Create cor_tbl extractor function}
\usage{
get_data(..., type = "full", show.diag = TRUE)
}
\arguments{
\item{...}{extra filter params, see Details.}

\item{type}{a string, "full" (default), "upper" or "lower", display full,
lower triangular or upper triangular matrix.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}
}
\value{
extractor function
}
\description{
This function returns another function that can extract cor_tbl
    subset from a cor_tbl object.
}
\details{
This function is mainly used in \code{ggplot2} geom_*() functions,
    where data is filtered based on the \code{...} parameter, then subsets
    are extracted based on the type and show.diag parameters.
}
\examples{
quickcor(mtcars) +
  geom_square(data = get_data(type = "upper")) +
  geom_circle2(data = get_data(type = "lower", show.diag = FALSE))
}
\seealso{
\code{\link[dplyr]{filter}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
