% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-order.R
\name{get_order}
\alias{get_order}
\alias{get_order.numeric}
\alias{get_order.character}
\alias{get_order.hclust}
\alias{get_order.dendrogram}
\title{Extract order}
\usage{
get_order(x, ...)

\method{get_order}{numeric}(x, ...)

\method{get_order}{character}(x, ...)

\method{get_order}{hclust}(x, ...)

\method{get_order}{dendrogram}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{ignore.}
}
\value{
a character or numeric vector.
}
\description{
Extract order
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
