% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcor-packages.R
\docType{package}
\name{ggcor_packages}
\alias{ggcor_packages}
\title{ggcor: Extended tools for correlation analysis and visualization}
\description{
The 'ggcor' package can be used to visualize simply and directly a
correlation matrix based on 'ggplot2'. It provides a solution for reordering the
correlation matrix, displaying the different significance level on the plot and
other details. The most important parts, It also provides a graphical display of
any correlation analysis and their combination (such as Mantel test, procrutes
test, and so on).
}
