% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liner-trans.R
\name{liner_trans}
\alias{liner_trans}
\title{Scale continuous}
\usage{
liner_trans(from, to)
}
\arguments{
\item{from}{input range.}

\item{to}{output range.}
}
\value{
liner trans.
}
\description{
Scale continuous with liner transform
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
