% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link-extra-params.R
\name{link_params}
\alias{link_params}
\title{Control the points position of link}
\usage{
link_params(
  env.point.hjust = NULL,
  env.point.vjust = NULL,
  spec.point.hjust = NULL,
  spec.point.vjust = NULL
)
}
\arguments{
\item{env.point.hjust, env.point.vjust}{a numeric vector is used to set the distance that
points (close to the correlation matrix) moves horizontally or vertically.}

\item{spec.point.hjust, spec.point.vjust}{a numeric vector is used to set the distance that
points (away from the correlation matrix) moves horizontally or vertically.}
}
\description{
This is mainly used in the add_link function to control points position.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
