% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-tbl.R
\name{link_tbl}
\alias{link_tbl}
\title{Transform data base on different layout}
\usage{
link_tbl(
  data,
  cor_tbl,
  start.var = NULL,
  start.name = NULL,
  end.var = NULL,
  pos = "right"
)
}
\arguments{
\item{data}{a data frame.}

\item{cor_tbl}{a col_tbl object.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
is forced to be converted.}

\item{start.name}{names of start point.}

\item{pos}{position of link.}
}
\value{
a data frame.
}
\description{
These layout functions are not layout in the network diagram,
it just converts the original data into a form that makes it easy to draw
a curve graph.
}
\examples{
\dontrun{
data(varespec, package = "vegan")
data(varechem, package = "vegan")
m <- mantel_test(varespec, varechem, spec.select = list(1:10, 11:44))
corr <- fortify_cor(varechem)
link_tbl(m, corr) ## full
link_tbl(m, get_upper_data(corr)) ## upper
link_tbl(m, get_lower_data(corr)) ## lower
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
