% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-network-statistics.R
\name{is_cor_network}
\alias{is_cor_network}
\alias{cn_active}
\alias{cn_degree}
\alias{cn_degree_distribution}
\alias{cn_diameter}
\alias{cn_distance_table}
\alias{cn_mean_distance}
\alias{cn_distances}
\alias{cn_shortest_paths}
\alias{cn_all_shortest_paths}
\alias{cn_estimate_betweenness}
\alias{cn_betweenness}
\alias{cn_edge_betweenness}
\alias{cn_edge_density}
\alias{cn_edge_connectivity}
\alias{cn_ego_size}
\alias{cn_ego}
\alias{cn_make_ego_graph}
\alias{cn_eccentricity}
\alias{cn_ecount}
\alias{cn_eigen_centrality}
\alias{cn_embed_adjacency_matrix}
\alias{cn_embed_laplacian_matrix}
\alias{cn_closeness}
\alias{cn_estimate_closeness}
\alias{cn_transitivity}
\alias{cn_centr_betw}
\alias{cn_centr_betw_tmax}
\alias{cn_centr_clo}
\alias{cn_centr_clo_tmax}
\alias{cn_centr_degree}
\alias{cn_centr_degree_tmax}
\alias{cn_centr_eigen}
\alias{cn_centr_eigen_tmax}
\alias{cn_component_distribution}
\alias{cn_components}
\alias{cn_cluster_edge_betweenness}
\alias{cn_cluster_fast_greedy}
\alias{cn_cluster_infomap}
\alias{cn_cluster_label_prop}
\alias{cn_cluster_leading_eigen}
\alias{cn_cluster_louvain}
\alias{cn_cluster_optimal}
\alias{cn_cluster_spinglass}
\alias{cn_cluster_walktrap}
\alias{cn_vertex_connectivity}
\alias{cn_cohesion}
\alias{cn_coreness}
\title{Network statistics function}
\usage{
is_cor_network(cn)

cn_active(cn, active = "nodes")

cn_degree(cn, ...)

cn_degree_distribution(cn, ...)

cn_diameter(cn, ...)

cn_distance_table(cn, ...)

cn_mean_distance(cn, ...)

cn_distances(cn, ...)

cn_shortest_paths(cn, ...)

cn_all_shortest_paths(cn, ...)

cn_estimate_betweenness(cn, ...)

cn_betweenness(cn, ...)

cn_edge_betweenness(cn, ...)

cn_edge_density(cn, ...)

cn_edge_connectivity(cn, ...)

cn_ego_size(cn, ...)

cn_ego(cn, ...)

cn_make_ego_graph(cn, ...)

cn_eccentricity(cn, ...)

cn_ecount(cn, ...)

cn_ecount(cn, ...)

cn_eigen_centrality(cn, ...)

cn_embed_adjacency_matrix(cn, ...)

cn_embed_laplacian_matrix(cn, ...)

cn_closeness(cn, ...)

cn_estimate_closeness(cn, ...)

cn_transitivity(cn, ...)

cn_centr_betw(cn, ...)

cn_centr_betw_tmax(cn, ...)

cn_centr_clo(cn, ...)

cn_centr_clo_tmax(cn, ...)

cn_centr_degree(cn, ...)

cn_centr_degree_tmax(cn, ...)

cn_centr_eigen(cn, ...)

cn_centr_eigen_tmax(cn, ...)

cn_component_distribution(cn, ...)

cn_components(cn, ...)

cn_cluster_edge_betweenness(cn, ...)

cn_cluster_fast_greedy(cn, ...)

cn_cluster_infomap(cn, ...)

cn_cluster_label_prop(cn, ...)

cn_cluster_leading_eigen(cn, ...)

cn_cluster_louvain(cn, ...)

cn_cluster_optimal(cn, ...)

cn_cluster_spinglass(cn, ...)

cn_cluster_walktrap(cn, ...)

cn_vertex_connectivity(cn, ...)

cn_cohesion(cn, ...)

cn_coreness(cn, ...)
}
\arguments{
\item{cn}{cor_network object.}

\item{...}{extra params.}
}
\description{
All network statistical functions are the encapsulation
    of corresponding functions in the igraph package. These function
    names start with "cn_" in ggcor, and the rest are corresponding
    function names in igraph.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
