% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-tbl-utils.R
\name{plot.cor_tbl}
\alias{plot.cor_tbl}
\title{Draw a cor_tbl object}
\usage{
\method{plot}{cor_tbl}(x, ...)
}
\arguments{
\item{x}{A cor_tbl object.}

\item{...}{Extra params passing to `ggcor()` function.}
}
\description{
Draw a cor_tbl object
}
\author{
Houyun Huang
}
