% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link-extra-params.R
\name{point_params}
\alias{point_params}
\title{Extra points params}
\usage{
point_params(
  alpha = NA,
  colour = "black",
  fill = NA,
  shape = 21,
  size = 1,
  stroke = 0.5,
  color = NULL
)
}
\arguments{
\item{alpha}{alpha channel for transparency.}

\item{colour, color}{colour of points.}

\item{fill}{fill colour of points.}

\item{shape}{shape of points.}

\item{size}{size of points.}

\item{stroke}{stroke of points.}
}
\description{
This is mainly used in the add_link function to control points style.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
