% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarcor.R
\name{polarcor}
\alias{polarcor}
\alias{geom_arc_colour}
\alias{set_arc_axis_x}
\alias{set_arc_axis_y}
\title{Visualize correlation matrix on polar coordinate}
\usage{
polarcor(data, no.axis = FALSE, ...)

geom_arc_colour(mapping = NULL, data = NULL, ...)

set_arc_axis_x(mapping = NULL, data = NULL, groups = NULL, ...)

set_arc_axis_y(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{data}{a data.frame.}

\item{no.axis}{if TRUE, don't add axis labels for plot.}

\item{...}{extra parameters.}

\item{mapping}{NULL (default) or a list of aesthetic mappings to use for plot.}

\item{groups}{character vectors.}
}
\description{
The \code{trans_polar} function transforms the cor_tbl
object into polar coordinate format, and \code{polarcor}, \code{geom_arc_colour}
and other functions are used for visualization.
}
\examples{
polarcor(mtcars, mapping = aes(fill = r))
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
