% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-shift.R
\docType{data}
\name{position_shift_dodge}
\alias{position_shift_dodge}
\alias{PositionShiftDodge}
\title{Dodge position with some shift}
\usage{
position_shift_dodge(
  width = NULL,
  xshift = NA,
  yshift = NA,
  preserve = c("total", "single")
)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{xshift, yshift}{the offsets on x-axis and y-axis.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}
}
\value{
a position object.
}
\description{
Shift dodge position function for annotation.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
\keyword{datasets}
