% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-shift.R
\docType{data}
\name{position_shift_dodge2}
\alias{position_shift_dodge2}
\alias{PositionShiftDodge2}
\title{Dodge2 position with some shift}
\usage{
position_shift_dodge2(
  width = NULL,
  preserve = c("total", "single"),
  xshift = NA,
  yshift = NA,
  padding = 0.1,
  reverse = FALSE
)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{xshift, yshift}{the offsets on x-axis and y-axis.}

\item{padding}{Padding between elements at the same position. Elements are
shrunk by this proportion to allow space between them. Defaults to 0.1.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\value{
a position object.
}
\description{
Shift dodge2 position function for annotation.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
\keyword{datasets}
