% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-shift.R
\docType{data}
\name{position_shift_identity}
\alias{position_shift_identity}
\alias{PositionShiftIdentity}
\title{identity position with some shift}
\usage{
position_shift_identity(xshift = NA, yshift = NA)
}
\arguments{
\item{xshift, yshift}{the offsets on x-axis and y-axis.}
}
\value{
a position object.
}
\description{
Shift identity position function for annotation.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
\keyword{datasets}
