% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-tbl-utils.R
\name{print.cor_tbl}
\alias{print.cor_tbl}
\title{Print a cor_tbl object}
\usage{
\method{print}{cor_tbl}(x, n = 5, width = 60, ...)
}
\arguments{
\item{x}{A cor_tbl object.}

\item{n}{The max lines will be print (the default is 5).}

\item{width}{The max width for each line.}

\item{...}{Extra params passing to `print()` function.}
}
\description{
Print a cor_tbl object
}
\author{
Houyun Huang
}
