% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-test.R
\name{print.correlation}
\alias{print.correlation}
\title{Print for correlate object.}
\usage{
\method{print}{correlation}(x, all = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{all}{if FALSE (default) just print correlation matrix, else will
print all values.}
}
\description{
Print for correlate object.
}
\examples{
m <- correlate(mtcars)
print(m)
print(m, TRUE)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
