% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrutes-test.R
\name{mono_mds}
\alias{mono_mds}
\alias{dudi_pca}
\title{Helper functions for procrutes test}
\usage{
mono_mds(x, method = "bray", ...)

dudi_pca(x, ...)
}
\arguments{
\item{x}{a data frame.}

\item{method}{the distance measure to be used.}

\item{...}{extra parameters}
}
\value{
a matrix.
}
\description{
\code{mono_mds} is used to transform data by \code{\link[vegan]{monoMDS}},
and \code{dudi_pca} is used to transform data by \code{\link[ade4]{dudi.pca}}.
}
\seealso{
\code{\link[vegan]{vegdist}}, \code{\link[vegan]{monoMDS}},
\code{\link[ade4]{dudi.pca}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei

Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
