% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/procrutes-test.R
\name{fortify_procrutes}
\alias{fortify_procrutes}
\alias{procrutes_test}
\title{Procrutes test for dissimilarity matrices}
\usage{
procrutes_test(
  spec,
  env,
  group = NULL,
  procrutes.fun = "protest",
  spec.select = NULL,
  env.select = NULL,
  use = "everything",
  spec.pre.fun = "identity",
  spec.pre.params = list(),
  env.pre.fun = spec.pre.fun,
  env.pre.params = spec.pre.params,
  seed = 123,
  ...
)
}
\arguments{
\item{...}{extra params passing to \code{procrutes.fun}.}

\item{spec, env}{data frame object.}

\item{group}{vector for rows grouping.}

\item{procrutes.fun}{string, name of procrutes test function.
 \itemize{
   \item{\code{"protest"} will use \code{vegan::protest()} (default).}
   \item{\code{"procuste.randtest"} will use \code{ade4::procuste.randtest()}.}
   \item{\code{"procuste.rtest"} will use \code{ade4::procuste.rtest()}.}
}}

\item{spec.select, env.select}{NULL (default), numeric or character vector index of columns.}

\item{use}{one of "everything", "complete" or "pairwise".}

\item{spec.pre.fun, env.pre.fun}{string, function name of transform the input data.}

\item{spec.pre.params, env.pre.params}{list, extra parameters for \code{spec/env.pre.fun}.}

\item{seed}{a integer value.}
}
\value{
a data frame.
}
\description{
Perform procrutes test quickly and tidy up the data to
data frame.
}
\examples{
\dontrun{
library(vegan)
data("varespec")
data("varechem")
procrutes_test(varespec, varechem)
procrutes_test(varespec, varechem, procrutes.fun = "procuste.randtest")
procrutes_test(varespec, varechem,
            spec.select = list(spec01 = 1:6, spec02 = 7:12))
procrutes_test(varespec, varechem, spec.pre.fun = "mono_mds",
            spec.select = list(spec01 = 1:6, spec02 = 7:12),
            env.select = list(env01 = 1:4, env02 = 5:14))
set.seed(20191224)
sam_grp <- sample(paste0("sample", 1:3), 24, replace = TRUE)
procrutes_test(varespec, varechem, group = sam_grp)
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
