% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qheatmap.R
\name{qheatmap}
\alias{qheatmap}
\title{Quick heatmap}
\usage{
qheatmap(data, mapping = NULL, geom = "tile", zoom_x = 1, zoom_y = 1, ...)
}
\arguments{
\item{data}{a cor_tbl object.}

\item{mapping}{aesthetic mappings parameters.}

\item{geom}{short name of layer function, the default is "tile".}

\item{zoom_x, zoom_y}{the ratio of zoom.}

\item{...}{extra parameters for layer function.}
}
\value{
a gg object.
}
\description{
Quick heatmap
}
\note{
This function is mainly used to add annotation.
}
\examples{
data <- gcor_tbl(mtcars)
qheatmap(data, aes(fill = value))

library(ggplot2)
p <- matrix(sample(LETTERS[1:3], 33, replace = TRUE), nrow = 11) \%>\%
     gcor_tbl() \%>\%
     qheatmap(aes(fill = value), zoom_x = 0.5) +
     remove_y_axis() +
     coord_fixed()
quickcor(mtcars) +
  geom_square() +
  anno_row_custom(p)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
