% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcor.R
\name{quickcor}
\alias{quickcor}
\title{Plot Correlation Matrix Quickly}
\usage{
quickcor(x, y = NULL, mapping = NULL, fill.colours = NULL,
  fill.bin = FALSE, panel.backgroud = NA, grid.colour = "grey50",
  grid.size = 0.25, grid.linetype = "solid",
  axis.x.position = "auto", axis.y.position = "auto",
  axis.label.drop = TRUE, legend.title = "corr",
  legend.position = "auto", legend.breaks = NULL,
  legend.labels = NULL, coord.fixed = TRUE, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x, y}{matrix or data frame.}

\item{is.cor}{logical value}
}
\description{
quickcor is convenient wrapper for creating a number of different types
of correlation matrix plots because of adding some extra settings by default.
}
