% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-correlate.R
\name{rand_correlate}
\alias{rand_correlate}
\alias{rand_dataset}
\title{Make dataset}
\usage{
rand_correlate(
  rows = 12,
  cols = NULL,
  obs = 100,
  row.names = NULL,
  col.names = NULL,
  type = "cubic",
  frequency = 0.15,
  reorder = TRUE,
  seed = NULL,
  ...
)

rand_dataset(
  vars = 12,
  obs = 100,
  reorder = TRUE,
  row.names = NULL,
  col.names = NULL,
  type = "cubic",
  frequency = 0.15,
  seed = NULL,
  ...
)
}
\arguments{
\item{rows, cols}{number of rows and columns.}

\item{obs}{number of observation.}

\item{row.names, col.names}{row/column names.}

\item{type}{one of "perlin", "cubic", "simplex", "value", "worley" or "white".}

\item{frequency}{determines the granularity of the features in the noise.}

\item{reorder}{reorder the dataset at random.}

\item{seed}{an integer value or NULL.}

\item{...}{extra parameters.}

\item{vars}{number of variables.}
}
\value{
a correlate object or data.frame.
}
\description{
Make rand dataset
}
\examples{
rand_dataset()
rand_correlate() \%>\% quickcor() + geom_square()
rand_correlate(cols = 6) \%>\% quickcor() + geom_square()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
