% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest.R
\name{rand_forest}
\alias{rand_forest}
\alias{print.rand_forest}
\title{Random forests}
\usage{
rand_forest(spec, env, use = "everything", byrow = FALSE, seed = 123, ...)

\method{print}{rand_forest}(x, ...)
}
\arguments{
\item{spec, env}{a data.frame object.}

\item{use}{one of "everything", "complete" or "pairwise", used to handle missing value.}

\item{byrow}{a logical value, if TRUE, the 'spec' on the rows.}

\item{seed}{a integer value.}

\item{...}{extra parameters.}

\item{x}{a rand_forest object.}
}
\value{
a rand_forest object.
}
\description{
Random forests
}
\examples{
\dontrun{
spec <- mtcars[c(1, 3, 4, 5)]
env <- mtcars[6:11]
rand_forest(spec, env)
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
