% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot_add}
\alias{aes}
\alias{\%>\%}
\alias{filter}
\alias{mutate}
\alias{group_by}
\alias{ungroup}
\alias{as_tibble}
\alias{as_tbl_graph}
\alias{as.igraph}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{ungroup}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{ggplot_add}}}

  \item{igraph}{\code{\link[igraph]{as.igraph}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidygraph}{\code{\link[tidygraph]{as_tbl_graph}}}
}}

