% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-axis.R
\name{remove_axis}
\alias{remove_axis}
\alias{remove_all_axis}
\alias{remove_x_axis}
\alias{remove_y_axis}
\title{Remove axis elements.}
\usage{
remove_axis(index = c("all", "x", "y"))

remove_all_axis()

remove_x_axis()

remove_y_axis()
}
\arguments{
\item{index}{'all' (default), 'x' or 'y' axis will be removed.}
}
\value{
The theme.
}
\description{
A simple wrapper of the \code{\link[ggplot2]{theme}} function
to quickly remove axis elements.
}
\examples{
quickcor(mtcars) + geom_circle2() + remove_axis()
quickcor(mtcars) + geom_circle2() + remove_axis("x")
quickcor(mtcars) + geom_circle2() + remove_axis("y")
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
