% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-cor-tbl.R
\name{reorder_cor_tbl}
\alias{reorder_cor_tbl}
\alias{get_order}
\alias{get_order.numeric}
\alias{get_order.character}
\alias{get_order.hclust}
\alias{get_order.dendrogram}
\title{Reorder cor_tbl}
\usage{
reorder_cor_tbl(
  x,
  row.order = NULL,
  col.order = if (is_symmet(x)) row.order else NULL
)

get_order(x, ...)

\method{get_order}{numeric}(x, ...)

\method{get_order}{character}(x, ...)

\method{get_order}{hclust}(x, ...)

\method{get_order}{dendrogram}(x, ...)
}
\arguments{
\item{x}{numeric vector, character vector, hclust object or dendrogram object.}

\item{row.order, col.order}{order of the correlation matrix, and it can be
NULL (default), numeric vector, character vector, hclust or dendrogram
object.}

\item{...}{parameters.}
}
\value{
a modified cor_tbl object.
}
\description{
Reorder cor_tbl object by name, hclut or or other method.
}
\examples{
fortify_cor(mtcars) \%>\% 
  reorder_cor_tbl(sample(1:11)) \%>\% 
  quickcor() + geom_square()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
