% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-style.R
\name{set_style}
\alias{set_style}
\alias{reset_style}
\title{Set style}
\usage{
set_style(
  colours = NULL,
  type = "gradient2n",
  legend.title = "Corr",
  nbin = 40,
  ...
)

reset_style()
}
\arguments{
\item{colours}{NULL or colours.}

\item{type}{a scale function or character of scale name (should be one of 
"gradient", "viridis" or "gradient2n").}

\item{legend.title}{a string of colour bar title.}

\item{nbin}{a scala integer.}

\item{...}{extra parameters for scale function.}
}
\description{
Set the style of correlation matrix plot
}
\examples{
set_style()
quickcor(mtcars) + geom_square()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
