% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link-extra-params.R
\name{text_params}
\alias{text_params}
\title{Extra text label params}
\usage{
text_params(colour = "black", size = 3.88, angle = 0, hjust = NULL,
  vjust = 0.5, alpha = NA, family = "", fontface = 1,
  color = NULL)
}
\arguments{
\item{colour, color}{colour of text.}

\item{size}{font size of text.}

\item{angle}{angle to rotate the text.}

\item{hjust, vjust}{a numeric vector specifying horizontal/vertical justification.}

\item{alpha}{alpha channel for transparency.}

\item{family}{the font family.}

\item{fontface}{the font face.}
}
\description{
This is mainly used in the add_link function to set the group label.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
