% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-cor.R
\name{theme_cor}
\alias{theme_cor}
\alias{theme_anno}
\alias{theme_anno2}
\title{Create the default ggcor theme}
\usage{
theme_cor(legend.position = "right", ...)

theme_anno()

theme_anno2()
}
\arguments{
\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector).}

\item{...}{extra params passing to \code{\link[ggplot2]{theme}}.}
}
\value{
The theme.
}
\description{
Create the default ggcor theme
}
\details{
The theme_cor, with no axis title, no background, no grid,
    made some adjustments to the x-axis label.
}
\examples{
require(ggplot2, quietly = TRUE)
df <- fortify_cor(mtcars)
ggcor(df) + geom_raster(aes(fill = r)) + theme_cor()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
