% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-network-data.R
\name{tidy_network_data}
\alias{tidy_network_data}
\title{Tidy co-occurrence network data}
\usage{
tidy_network_data(x, r.thres = 0.6, r.absolute = TRUE,
  p.thres = NULL, type = "upper", show.diag = FALSE,
  cor.test = TRUE, ...)
}
\arguments{
\item{x}{any \code{R} object can be converted to cor_tbl.}

\item{r.thres}{a numeric value.}

\item{r.absolute}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}

\item{...}{passing to \code{\link[ggcor]{as_cor_tbl}}}
}
\value{
cor_tbl.
}
\description{
The function calculates correlation coefficient, statistical
    significance level and filters according to conditions.
}
