% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarcor.R
\name{trans_polar}
\alias{trans_polar}
\title{Transforms cor_tbl into polar coordinate format}
\usage{
trans_polar(
  data,
  start.angle = NULL,
  end.angle = NULL,
  inner.radius = 0.4,
  outer.radius = 1,
  axis.x.radius = 1.02,
  split.by.group = FALSE,
  group.space = 1,
  ...
)
}
\arguments{
\item{data}{a cor_tbl object or others that can be convert to cor_tbl object.}

\item{start.angle, end.angle}{angle of start and end point (in degree).}

\item{inner.radius, outer.radius, axis.x.radius}{inner/outer/x axis labels radius.}

\item{split.by.group}{if TRUE, will split the plot region on groups.}

\item{group.space}{the space between each group (in degree).}

\item{...}{extra parameters.}
}
\value{
a data.frame.
}
\description{
\code{trans_polar} function transforms the cor_tbl
object into polar coordinate format, and this format can draw polar
matrix heatmap on cartesian coordinate.
}
\examples{
trans_polar(mtcars)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
