% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_hc_rect}
\alias{anno_hc_rect}
\title{Square annotation}
\usage{
anno_hc_rect(k = 2, fill = NA, colour = "black", size = 2, color)
}
\arguments{
\item{k}{an integer scalar or vector with the desired number of groups.}

\item{fill}{NA (default) or the fill colour of square.}

\item{colour, color}{the colour of square boder.}

\item{size}{size of square boder line.}
}
\value{
square layer.
}
\description{
Draw the cluster square mark on the correlation matrix plot.
}
\details{
Square annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
