% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_link}
\alias{anno_link}
\alias{anno_link_label}
\title{Special annotation function for correlation link plot}
\usage{
anno_link(
  mapping = NULL,
  data,
  start.var = NULL,
  start.name = NULL,
  end.var = NULL,
  ...
)

anno_link_label(
  mapping = NULL,
  nudge_x = 0.1,
  geom = "text",
  is.start = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{a data frame.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
will be converted.}

\item{...}{extra parameters passing to layer function.}

\item{nudge_x}{horizontal adjustment to nudge labels by.}

\item{geom}{one of "text", "label" or "image".}

\item{is.start}{NULL (default), TRUE or FALSE.}
}
\value{
a ggplot layer.
}
\description{
a set of custom layer functions that quickly add
layers of curves, nodes, and labels.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
