% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_link}
\alias{anno_link}
\title{Special annotation function for correlation link plot}
\usage{
anno_link(
  mapping = NULL,
  data,
  width = 0.3,
  pos = "right",
  label.size = 3.5,
  label.colour = "black",
  label.family = "",
  label.fontface = "",
  nudge_x = 0.1,
  expand = NULL,
  start.var = NULL,
  start.name = NULL,
  end.var = NULL,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{a data frame.}

\item{width}{width of link plot.}

\item{pos}{position of link plot.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{nudge_x}{horizonal justification of label.}

\item{expand}{expansion of x axis.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
will be converted.}

\item{...}{extra parameters passing to layer function.}
}
\value{
a ggplot layer.
}
\description{
a set of custom layer functions that quickly add
layers of curves, nodes, and labels.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
