% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_bar}
\alias{anno_bar}
\alias{anno_bar2}
\alias{anno_boxplot}
\alias{anno_point}
\alias{anno_row_heat}
\alias{anno_col_heat}
\title{Custom annotation}
\usage{
anno_bar(
  mapping,
  data,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_bar2(
  mapping,
  data,
  align = TRUE,
  flip = FALSE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_boxplot(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_point(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_row_heat(
  mapping,
  data,
  align = TRUE,
  geom = "anno_tile",
  mark = NULL,
  space = 0.5,
  width = 1,
  col.label = TRUE,
  ...
)

anno_col_heat(
  mapping,
  data,
  align = TRUE,
  geom = "anno_tile2",
  mark = NULL,
  space = 0.5,
  height = 1,
  row.label = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{a data frame.}

\item{align}{align base on main plot.}

\item{flip}{a logical value. If TRUE, the annotation on the left or on
the bottom side will be flipped.}

\item{pos}{position of annotation.}

\item{width, height}{width and height of annotation.}

\item{height}{scala numeric value.}

\item{trans}{the name of a transformation object or the object itself.}

\item{remove.axis}{one of "auto", "x", "y" or "none".}

\item{theme}{theme object or the name.}

\item{...}{extra parameters.}

\item{geom}{one of "anno_tile", "anno_tile2" or "point".}

\item{space}{scala numeric value.}

\item{col.label}{logical value, if TRUE will add columns name on plot.}

\item{row.label}{logical value, if TRUE will add columns name on plot.}
}
\value{
anno_* object.
}
\description{
Custom annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
