% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_bar}
\alias{anno_bar}
\alias{anno_bar2}
\alias{anno_boxplot}
\alias{anno_point}
\title{Custom annotation}
\usage{
anno_bar(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_bar2(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_boxplot(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  trans = NULL,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)

anno_point(
  mapping,
  data,
  align = TRUE,
  pos = NULL,
  width = 0.2,
  height = 0.2,
  remove.axis = "auto",
  theme = theme_anno(),
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{a data frame.}

\item{align}{align base on main plot.}

\item{pos}{position of annotation.}

\item{width, height}{width and height of annotation.}

\item{trans}{the name of a transformation object or the object itself.}

\item{remove.axis}{one of "auto", "x", "y" or "none".}

\item{theme}{theme object or the name.}

\item{...}{extra parameters.}
}
\value{
anno_* object.
}
\description{
Custom annotation
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
