% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_tree}
\alias{anno_tree}
\title{Annotation for correlation matrix plot}
\usage{
anno_tree(index = "all", colour = NULL, size = NULL, linetype = NULL, color)
}
\arguments{
\item{index}{one of "all", "row" or "col".}

\item{colour, color}{colour of segments.}

\item{size}{width of segments.}

\item{linetype}{line type of segments.}
}
\value{
anno_tree object.
}
\description{
Annotation for correlation matrix plot
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
