% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cor-network.R
\name{as_cor_network}
\alias{as_cor_network}
\alias{as_cor_network.cor_tbl}
\alias{as_cor_network.default}
\title{Tidy co-occurrence network data}
\usage{
as_cor_network(x, ...)

\method{as_cor_network}{cor_tbl}(x, simplify = TRUE, r.thres = 0.6,
  r.absolute = TRUE, p.thres = 0.05, ...)

\method{as_cor_network}{default}(x, simplify = TRUE, r.thres = 0.6,
  r.absolute = TRUE, p.thres = 0.05, type = "upper", ...)
}
\arguments{
\item{x}{any \code{R} object can be converted to cor_tbl.}

\item{...}{passing to \code{\link[ggcor]{fortify_cor}}}

\item{simplify}{logical value (defaults to TRUE) indicating whether to
delete nodes without edge connections.}

\item{r.thres}{a numeric value.}

\item{r.absolute}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}

\item{type}{a string, "full" (default), "upper" or "lower", display full,
lower triangular or upper triangular matrix.}
}
\value{
a list of nodes and edges.
}
\description{
The function calculates correlation coefficient, statistical
    significance level and filters according to conditions.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
