% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cor-network.R
\name{as_cor_network}
\alias{as_cor_network}
\alias{as_cor_network.cor_tbl}
\alias{as_cor_network.mantel_tbl}
\alias{as_cor_network.matrix}
\alias{as_cor_network.data.frame}
\alias{as_cor_network.correlate}
\alias{as_cor_network.rcorr}
\alias{as_cor_network.corr.test}
\alias{as_cor_network.igraph}
\alias{as_cor_network.tbl_graph}
\alias{as_cor_network.default}
\title{Coerce to a cor_network object}
\usage{
as_cor_network(x, ...)

\method{as_cor_network}{cor_tbl}(x, simplify = TRUE, r.thres = 0.6,
  r.absolute = TRUE, p.thres = 0.05, ...)

\method{as_cor_network}{mantel_tbl}(x, ...)

\method{as_cor_network}{matrix}(x, ...)

\method{as_cor_network}{data.frame}(x, ...)

\method{as_cor_network}{correlate}(x, ...)

\method{as_cor_network}{rcorr}(x, ...)

\method{as_cor_network}{corr.test}(x, ...)

\method{as_cor_network}{igraph}(x, ...)

\method{as_cor_network}{tbl_graph}(x, ...)

\method{as_cor_network}{default}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{extra params passing to \code{\link[ggcor]{cor_network}}.}

\item{simplify}{logical value (defaults to TRUE) indicating whether to
delete nodes without edge connections.}

\item{r.thres}{a numeric value.}

\item{r.absolute}{logical value (defaults to TRUE).}

\item{p.thres}{a numeric value.}
}
\value{
a cor_network object.
}
\description{
Functions to coerce a object to cor_network if possible.
}
\examples{
ll <- correlate(mtcars)
as_cor_network(ll)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
