% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cor-tbl.R
\name{as_cor_tbl}
\alias{as_cor_tbl}
\alias{as_cor_tbl.list}
\alias{as_cor_tbl.matrix}
\alias{as_cor_tbl.data.frame}
\alias{as_cor_tbl.correlation}
\alias{as_cor_tbl.rcorr}
\alias{as_cor_tbl.corr.test}
\alias{as_cor_tbl.mantel_tbl}
\alias{as_cor_tbl.default}
\title{Coerce to a cor_tbl object.}
\usage{
as_cor_tbl(corr, ...)

\method{as_cor_tbl}{list}(x, type = "full", show.diag = TRUE,
  row.names = NULL, col.names = NULL, cluster = FALSE, ...)

\method{as_cor_tbl}{matrix}(corr, ...)

\method{as_cor_tbl}{data.frame}(corr, ...)

\method{as_cor_tbl}{correlation}(corr, ...)

\method{as_cor_tbl}{rcorr}(corr, check = FALSE, ...)

\method{as_cor_tbl}{corr.test}(corr, check = FALSE, ...)

\method{as_cor_tbl}{mantel_tbl}(corr, byrow = TRUE, ...)

\method{as_cor_tbl}{default}(corr, ...)
}
\arguments{
\item{corr}{any \code{R} object.}

\item{...}{extra params passing to \code{\link[ggcor]{matrix_order}}.}

\item{type}{a string, "full" (default), "upper" or "lower", display full,
lower triangular or upper triangular matrix.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}

\item{cluster}{a logical value indicating whether reorder the correlation matrix
by clustering, default is FALSE.}

\item{byrow}{a logical value indicating whether arrange the 'spec' columns on y axis.}

\item{rownames, colnames}{row/column names of correlation matrix.}
}
\value{
a cor_tbl object.
}
\description{
Functions to coerce a object to cor_tbl if possible.
}
\details{
\code{cluster = TRUE} just supports for symmetric correlation matrix.
}
\examples{
corr <- cor(mtcars)
as_cor_tbl(corr)
ll <- correlate(mtcars)
as_cor_tbl(ll, type = "upper")
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
