% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-matrix.R
\name{as_matrix}
\alias{as_matrix}
\alias{as_matrix.cor_tbl}
\alias{as_matrix.mantel_tbl}
\title{Convert a object to matrix}
\usage{
as_matrix(x, ...)

\method{as_matrix}{cor_tbl}(x, index = "all", missing = NULL, ...)

\method{as_matrix}{mantel_tbl}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{extra params.}

\item{index}{character vector indicating which columns will be convert. If "all",
all columns will be convert.}

\item{missing}{If NULL (default), the missing value will be filled with NAs.}
}
\value{
a list of matrix.
}
\description{
Functions to convert cor_tbl object to a list of matrix.
}
\examples{
cor(mtcars) \%>\% as_cor_tbl() \%>\% as_matrix()
cor(mtcars) \%>\% as_cor_tbl() \%>\% as_matrix("r")
fortify_cor(iris[-5],group = iris$Species, cor.test = TRUE) \%>\%
  as_matrix()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
