% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tbl-graph.R
\name{as_tbl_graph.cor_tbl}
\alias{as_tbl_graph.cor_tbl}
\alias{as_tbl_graph.mantel_tbl}
\alias{as_tbl_graph.pro_tbl}
\alias{as_tbl_graph.rcorr}
\alias{as_tbl_graph.corr.test}
\alias{as_tbl_graph.correlate}
\alias{as_tbl_graph.cor_network}
\title{Corece to a graph_tbl object}
\usage{
\method{as_tbl_graph}{cor_tbl}(x, directed = FALSE, ...)

\method{as_tbl_graph}{mantel_tbl}(x, directed = FALSE, ...)

\method{as_tbl_graph}{pro_tbl}(x, directed = FALSE, ...)

\method{as_tbl_graph}{rcorr}(x, directed = FALSE, ...)

\method{as_tbl_graph}{corr.test}(x, directed = FALSE, ...)

\method{as_tbl_graph}{correlate}(x, directed = FALSE, ...)

\method{as_tbl_graph}{cor_network}(x, ...)
}
\arguments{
\item{x}{\code{R} object.}

\item{directed}{logical value, whether or not to create a directed graph.}

\item{...}{extra params.}
}
\value{
tbl_graph object.
}
\description{
Functions to coerce a object to graph_tbl if possible.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
