% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-tbl-utils.R
\name{cor_tbl_xname}
\alias{cor_tbl_xname}
\alias{cor_tbl_yname}
\alias{cor_tbl_type}
\alias{cor_tbl_showdiag}
\title{Get attributes of cor_tbl}
\usage{
cor_tbl_xname(x)

cor_tbl_yname(x)

cor_tbl_type(x)

cor_tbl_showdiag(x)
}
\arguments{
\item{x}{a cor_tbl.}
}
\value{
return attribute vector.
}
\description{
Helper function to get extra attributs of cor_tbl.
}
\examples{
df <- fortify_cor(mtcars)
## get xname
cor_tbl_xname(df)
## get yname
cor_tbl_yname(df)
## get show.diag
cor_tbl_showdiag(df)
## get type
cor_tbl_type(df)
}
\author{
Houyun Huang
}
