% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-test.R
\name{correlate}
\alias{correlate}
\alias{fast_correlate}
\title{Matrix of Correlations, P-values and confidence intervals}
\usage{
correlate(x, y = NULL, cor.test = FALSE, method = "pearson",
  use = "everything", ...)

fast_correlate(x, y = NULL, use = "everything", ...)
}
\arguments{
\item{x, y}{a matrix object or NULL.}

\item{cor.test}{logical, if \code{TRUE} (default) will test for correlation.}

\item{method}{a character string indicating which correlation coefficient is to be used
for the test. One of "pearson", "kendall", or "spearman".}

\item{use}{an optional character string giving a method for computing covariances in the presence of missing values.}

\item{...}{extra params passing to \code{cor.test} for \code{correlate},
passing to \code{WGCNA::corAndPvalue}.}
}
\value{
a list with correlation matrix, P values matrix, confidence intervals matrix.
}
\description{
\code{correlate} uses \code{cor} to find the correlations and use \code{cor.test} to find
    the p values, confidence intervals for all possible pairs of columns ofmatrix.
}
\details{
The columns of 'x' will be tested for each pair when y is NULL(the default),
    otherwise each column in 'x' and each column in 'y' is tested for each pair.
}
\examples{
correlate(mtcars)
m1 <- matrix(rnorm(100), nrow = 10)
m2 <- matrix(rnorm(60), nrow = 10)

## not test for correlation matrix
correlate(m1, m2)

## test for correlation matrix
correlate(m1, m2, cor.test = TRUE)

## fast compute correlation
if(require(WGCNA)) {
  fast_correlate(m1, m2)
}
}
\seealso{
\code{\link[stats]{cor}}, \code{\link[stats]{cor.test}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
