% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-layout.R
\name{parallel_layout}
\alias{parallel_layout}
\alias{triangle_layout}
\alias{add_link}
\title{Transform data base on different layout}
\usage{
parallel_layout(
  data,
  type = NULL,
  show.diag = NULL,
  row.names = NULL,
  col.names = NULL,
  start.var = NULL,
  end.var = NULL,
  stretch = FALSE,
  cor_tbl
)

triangle_layout(
  data,
  type = NULL,
  show.diag = NULL,
  row.names = NULL,
  col.names = NULL,
  start.var = NULL,
  end.var = NULL,
  cor_tbl
)
}
\arguments{
\item{data}{a data frame.}

\item{type}{the type (""upper" or "lower") of the correlation matrix plot.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}

\item{row.names, col.names}{row/column names of correlation matrix.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
is forced to be converted.}

\item{stretch}{logical, indicating whether the heights/width of start points and
end points are consistent.}

\item{cor_tbl}{a col_tbl object.}
}
\value{
a data frame.
}
\description{
These layout functions are not layout in the network diagram,
it just converts the original data into a form that makes it easy to draw
a curve graph.
}
\examples{
cor_tbl(cor(mtcars)) \%>\%
  parallel_layout()
\dontrun{
data("varespec", package = "vegan")
data("varechem", package = "vegan")
mantel_test(varespec, varechem) \%>\%
  combination_layout(type = "upper", col.names = colnames(varechem),
                     show.diag = FALSE)
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
