% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot-utils.R
\name{as_dplot}
\alias{as_dplot}
\alias{is_dplot}
\alias{anno_col_custom}
\alias{anno_row_custom}
\alias{empty_plot}
\alias{print.dplot}
\title{Decoration plot}
\usage{
as_dplot(plot)

is_dplot(plot)

anno_col_custom(obj, height = 0.2, pos = "top")

anno_row_custom(obj, width = 0.2, pos = "right")

empty_plot()

\method{print}{dplot}(
  x,
  colours = getOption("ggcor.fill.pal"),
  style = getOption("ggcor.plot.style", "corrplot"),
  title = "corr",
  breaks = c(-1, -0.5, 0, 0.5, 1),
  labels = c(-1, -0.5, 0, 0.5, 1),
  limits = c(-1, 1),
  nbin = 40,
  ...
)
}
\arguments{
\item{plot, x}{a "gg" or "dplot" object.}

\item{obj}{a "gg" object.}

\item{pos}{one of "left", "right", "bottom" or "top".}

\item{width, height}{scala numeric value.}

\item{colours}{colour palette for filling.}

\item{style}{style of plot, one of "corrplot" (default) or "ggplot".}

\item{title}{guide title.}

\item{breaks}{breaks of guide_colourbar.}

\item{labels}{labels of guide_colourbar.}

\item{limits}{limits of guide_colourbar.}

\item{nbin}{a numeric specifying the number of bins for drawing the guide_colourbar.}

\item{...}{extra parameters.}
}
\value{
a "dplot" object.
}
\description{
Decoration plot
}
\examples{
\dontrun{
library(ggplot2)
d <- matrix(sample(c(LETTERS[1:3], NA), 33, replace = TRUE), nrow = 11)
p <- gcor_tbl(d, "var") \%>\%
  quickcor(mapping = aes(colour = var)) +
  geom_point(size = 3)
quickcor(mtcars) +
  geom_square() +
  anno_row_custom(p)
}
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
