% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-link-extra-params.R
\name{extra_params}
\alias{extra_params}
\title{Extra params for add_link}
\usage{
extra_params(
  spec.label = text_params(),
  spec.point = point_params(fill = "blue"),
  env.point = point_params(fill = "grey60"),
  link.params = link_params()
)
}
\arguments{
\item{spec.label}{NULL or "text_params" object producing by
\code{\link[ggcor]{text_params}}.}

\item{spec.point}{NULL or "point_params" object producing by
\code{\link[ggcor]{point_params}}.}

\item{env.point}{NULL or "point_params" object producing by
\code{\link[ggcor]{point_params}}.}

\item{link.params}{"point_params" object producing by \code{\link[ggcor]{point_params}}.}
}
\description{
This function is used to control the details
    of the link, including the location, shape, size, color
    of points, and font size, font color, and so on of the
    text label.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
