% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_lower_data}
\alias{get_lower_data}
\alias{get_upper_data}
\alias{get_diag_tri}
\alias{get_diag_data}
\title{Helper function to extract cor_tbl.}
\usage{
get_lower_data(x, show.diag = TRUE)

get_upper_data(x, show.diag = TRUE)

get_diag_tri(x)

get_diag_data(x)
}
\arguments{
\item{x}{a cor_tbl object.}

\item{show.diag}{a logical value indicating whether keep the diagonal.}
}
\value{
a modified cor_tbl object.
}
\description{
These functions are used to quickly obtain the upper
    trig, lower trig, diagonal, or remove the diagonal of the correlation
    coefficient matrix.
}
\examples{
df <- fortify_cor(mtcars)
quickcor(df) + geom_colour()
df01 <- get_lower_data(df)
quickcor(df01) + geom_colour()
get_upper_data(df, show.diag = FALSE)
quickcor(df02) + geom_colour()
df03 <- get_diag_data(df)
quickcor(df03) + geom_colour()
df04 <- get_diag_tri(df)
quickcor(df04) + geom_colour()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
