% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-colour.R
\name{geom_colour}
\alias{geom_colour}
\title{Unit square based on center}
\usage{
geom_colour(mapping = NULL, data = NULL, colour = "grey60", ...)
}
\arguments{
\item{mapping}{set of aesthetic mappings created by aes() or aes_().}

\item{data}{the data to be displayed in this layer.}

\item{colour}{the border colour.}

\item{...}{other params pass to `geom_tile()`.}
}
\description{
This is the encapsulation of \code{geom_tile()}
function, and some fine-tuning has been done.
}
\examples{
ggcorr(mtcars) + geom_colour()
}
