% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-diag-label.R
\name{geom_diag_label}
\alias{geom_diag_label}
\alias{add_diag_label}
\title{Add diagnoal labels on correlation plot}
\usage{
geom_diag_label(mapping = NULL, data = NULL, drop = FALSE, ...)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{data}{NULL (default) or a cor_tbl object.}

\item{drop}{logical value (default is TRUE). When type of plot is 'upper'
or 'lower' and 'show.diag' is FALSE, whether need to remove the blank label.}

\item{...}{extra params for \code{\link[ggplot2]{geom_text}}.}
}
\description{
\code{geom_diag_label} is mainly used with \code{ggcor} and
    \code{quickcor} functions to add diagnoal labels on correct position
    base on different type of cor_tbl object.
}
\examples{
quickcor(mtcars, type = "upper") + geom_colour() + geom_diag_label()
quickcor(mtcars, type = "lower") + geom_colour() + geom_diag_label()
}
\seealso{
\code{\link[ggplot2]{geom_text}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
