% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-diag-label.R, R/geom-grid.R
\name{geom_diag_label}
\alias{geom_diag_label}
\alias{add_diag_label}
\alias{add_grid}
\title{Add dianoal labels on correlation plot}
\usage{
geom_diag_label(data = NULL, drop = FALSE, ...)

add_diag_label(...)

add_grid(...)
}
\arguments{
\item{data}{NULL (default) or a cor_tbl object.}

\item{drop}{logical value (default is TRUE). When type of plot is 'upper'
or 'lower' and 'show.diag' is FALSE, whether need to remove the blank label.}

\item{...}{extra params for \code{\link[ggplot2]{geom_text}}.}
}
\description{
\code{add_grid} is mainly used with \code{ggcor} and
    \code{quickcor} functions.
}
\examples{
quickcor(mtcars, type = "upper") + geom_colour() + geom_diag_label()
quickcor(mtcars, type = "lower") + geom_colour() + geom_diag_label()
}
\seealso{
\code{\link[ggplot2]{geom_text}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
