% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-link.R, R/geom-link2.R
\docType{data}
\name{geom_links}
\alias{geom_links}
\alias{geom_links_label}
\alias{GeomLinks2}
\title{Special layer function for correlation link plot}
\usage{
geom_links(
  mapping = NULL,
  data,
  curvature = 0,
  layout = NULL,
  layout.params = list(),
  ...
)

geom_links_label(
  mapping = NULL,
  nudge_x = 0.1,
  geom = "text",
  is.start = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mappings parameters.}

\item{curvature}{a numeric value giving the amount of curvature.}

\item{layout}{one of "triangle" or "parallel".}

\item{layout.params}{parameters passing to layout function.}

\item{...}{extra parameters passing to layer function.}

\item{nudge_x}{horizontal adjustment to nudge labels by.}

\item{is.start}{NULL (default), TRUE or FALSE.}
}
\value{
a ggplot layer.
}
\description{
A set of custom layer functions that quickly add
layers of curves, nodes, and labels.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
\keyword{datasets}
