% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/geom-panel-grid.R
\name{add_grid}
\alias{add_grid}
\alias{geom_panel_grid}
\title{Add panel grid line on correlation plot}
\usage{
geom_panel_grid(colour = "grey50", size = 0.25, color = NULL)
}
\arguments{
\item{colour, color}{colour of grid lines.}

\item{size}{size of grid lines.}
}
\description{
\code{geom_grid} is mainly used with \code{ggcor} or \code{quickcor}
    function to add a panel grid line on plot region.
}
\examples{
df <- fortify_cor(mtcars)
ggcor(df) + geom_panel_grid()
require(ggplot2, quietly = TRUE)
ggplot(df, aes(x, y)) + geom_panel_grid()
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
