% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-order.R
\name{get_order}
\alias{get_order}
\alias{get_order.numeric}
\alias{get_order.character}
\alias{get_order.hclust}
\alias{get_order.dendrogram}
\alias{get_order.ggtree}
\alias{get_order.cor_tbl}
\alias{get_order.quickcor}
\title{Extract order}
\usage{
get_order(x, ...)

\method{get_order}{numeric}(x, ...)

\method{get_order}{character}(x, name = NULL, ...)

\method{get_order}{hclust}(x, ...)

\method{get_order}{dendrogram}(x, ...)

\method{get_order}{ggtree}(x, name = NULL, ...)

\method{get_order}{cor_tbl}(x, index = "row", name = NULL, ...)

\method{get_order}{quickcor}(x, index = "row", name = NULL, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{ignore.}

\item{name}{old name.}

\item{index}{one of "row" or "column".}
}
\value{
a numeric vector.
}
\description{
Extract order
}
\examples{
hc <- hclust(dist(mtcars))
get_order(hc)
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
