% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous-liner.R
\name{scale_x_liner}
\alias{scale_x_liner}
\title{Scale continuous}
\usage{
scale_x_liner(from, to, ...)
}
\arguments{
\item{from}{input range.}

\item{to}{output range.}

\item{...}{extra parameters.}
}
\value{
liner trans.
}
\description{
Scale continuous with liner transform
}
\examples{
\dontrun{
library(ggplot2)
df <- data.frame(x = sample(letters[1:10], 200, replace = TRUE))
p <- ggplot(df, aes(x)) + geom_bar()
from <- c(0, max(table(df$x)))
p + scale_y_liner(from, c(0, 3)) + coord_fixed()
}
}
