% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-layout.R
\name{link_tbl}
\alias{link_tbl}
\alias{add_link}
\title{Transform data base on different layout}
\usage{
link_tbl(data, cor_tbl, start.var = NULL, end.var = NULL, stretch = FALSE)
}
\arguments{
\item{data}{a data frame.}

\item{cor_tbl}{a col_tbl object.}

\item{start.var, end.var}{character to specify which variable is the starting
points and which is the ending points. if the variable is not character, it
is forced to be converted.}
}
\value{
a data frame.
}
\description{
These layout functions are not layout in the network diagram,
it just converts the original data into a form that makes it easy to draw
a curve graph.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
