% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-mantel.R
\name{mantel_test}
\alias{mantel_test}
\title{Mantel and partial mantel test for dissimilarity matrices}
\usage{
mantel_test(spec, env, env.ctrl = NULL, mantel.fun = "mantel",
  spec.select = NULL, env.select = NULL, spec.dist.method = "bray",
  env.dist.method = "euclidean", ...)
}
\arguments{
\item{spec, env}{data frame object.}

\item{env.ctrl}{NULL (default), data frame.}

\item{mantel.fun}{string, function of mantel test.
 \itemize{
   \item{\code{"mantel"} will use \code{vegan::mantel} (default).}
   \item{\code{"mantel.randtest"} will use \code{ade4::mantel.randtest}.}
   \item{\code{"mantel.rtest"} will use \code{ade4::mantel.rtest}.}
   \item{\code{"mantel.partial"} will use \code{vegan::mantel.partial} (default).}
}}

\item{spec.select, env.select}{NULL (default), numeric or character vector index of columns.}

\item{spec.dist.method}{dissimilarity index (default is 'bray'), pass to \code{method}
params of \code{\link[vegan]{vegdist}}.}

\item{env.dist.method}{dissimilarity index (default is euclidean'), pass to \code{method}
params of \code{\link[vegan]{vegdist}}.}

\item{...}{extra params for \code{mantel.fun}.}
}
\description{
Perform mantel test quickly and tidy up the data to
    data frame.
}
\examples{
library(vegan)
data("varespec")
data("varechem")
mantel_test(varespec, varechem)
mantel_test(varespec, varechem, mantel.fun = "mantel.randtest")
mantel_test(varespec, varechem, mantel.fun = "mantel.randtest",
  spec.select = list(spec01 = 1:6, spec02 = 7:12))
mantel_test(varespec, varechem, mantel.fun = "mantel.randtest",
  spec.select = list(spec01 = 1:6, spec02 = 7:12),
  env.select = list(env01 = 1:4, env02 = 5:14))
nm <- names(varechem[1:9])
mantel_test(varespec, varechem, env.ctrl = varechem[10:14],
  mantel.fun = "mantel.partial",
  env.select = as.list(setNames(nm, nm)))
}
\seealso{
\code{\link[vegan]{vegdist}}, \code{\link[vegan]{mantel}},
    \code{\link[ade4]{mantel.rtest}}, \code{\link[ade4]{mantel.randtest}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
