% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-order.R
\name{matrix_order}
\alias{matrix_order}
\title{Reorder Matrices}
\usage{
matrix_order(x, cluster.type = c("none", "all", "row", "col"),
  cluster.method = "HC", absolute = FALSE, ...)
}
\arguments{
\item{x}{a matrix object.}

\item{cluster.type}{a character string with the name of reorder type.}

\item{cluster.method}{a character string with the name of the seriation method.}

\item{absolute}{logical, if TRUE will transform x using abs().}

\item{...}{extra params passing to \code{\link[seriation]{seriate}}.}
}
\value{
a list of two numeric vector. The 'row_ord' is the order of rows, and
    the 'col_ord' is the order of columns.
}
\description{
Tries to find an linear order for matrix by different cluster methods.
}
\details{
All cluster methods in \code{\link[seriation]{seriate}} are supported，and the
    default cluster method is "HC". other methods please see `?seriation::seriate`.
}
\examples{
m <- matrix(rnorm(20), nrow = 4)
matrix_order(m, cluster.type = "all")
}
\seealso{
\code{\link[seriation]{seriate}}.
}
\author{
Houyun Huang

Lei Zhou

Jian Chen

Taiyun Wei
}
