% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-order.R
\name{matrix_order}
\alias{matrix_order}
\title{Reorder Matrices}
\usage{
matrix_order(x, is.cor = TRUE, cluster.method = "complete", ...)
}
\arguments{
\item{x}{a matrix-like object.}

\item{is.cor}{logical value (defaults to TRUE) indicating wheater
\code{x} is a correlation matrix.}

\item{cluster.method}{a character string with the name of agglomeration method.}

\item{...}{extra params passing to \code{\link[stats]{hclust}}.}
}
\value{
a list of hclust object.
}
\description{
Tries to find an order for matrix by different cluster methods.
}
\details{
Now it just supports for square matrix.
}
\examples{
m <- matrix(rnorm(25), nrow = 5)
matrix_order(m)
}
\seealso{
\code{\link[stats]{hclust}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
