% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-shift.R
\docType{data}
\name{position_shift_stack}
\alias{position_shift_stack}
\alias{PositionShiftStack}
\title{Pisition function with some shift and re-scale}
\usage{
position_shift_stack(
  vjust = 1,
  xshift = NA,
  yshift = NA,
  xscale = NULL,
  yscale = NULL,
  reverse = FALSE
)
}
\arguments{
\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}

\item{xshift, yshift}{the offsets on x-axis and y-axis.}

\item{xscale, yscale}{the range of the charts.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\value{
a position object.
}
\description{
Shift stack position function for annotation.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
\keyword{datasets}
