% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrutes-test.R
\name{procrutes_test}
\alias{procrutes_test}
\title{Procrutes test for dissimilarity matrices}
\usage{
procrutes_test(
  spec,
  env,
  procrutes.fun = "protest",
  spec.select = NULL,
  env.select = NULL,
  spec.pre.fun = "identity",
  spec.pre.params = list(),
  env.pre.fun = spec.pre.fun,
  env.pre.params = spec.pre.params,
  ...
)
}
\arguments{
\item{spec, env}{data frame object.}

\item{procrutes.fun}{string, name of procrutes test function.
 \itemize{
   \item{\code{"protest"} will use \code{vegan::protest} (default).}
   \item{\code{"procuste.randtest"} will use \code{ade4::procuste.randtest}.}
   \item{\code{"procuste.rtest"} will use \code{ade4::procuste.rtest}.}
}}

\item{spec.select, env.select}{NULL (default), numeric or character vector index of columns.}

\item{spec.pre.fun, env.pre.fun}{string, function name of transform the input data.}

\item{spec.pre.params, env.pre.params}{list, extra parameters for \code{spec/env.pre.fun}.}

\item{...}{extra params for \code{procrutes.fun}.}
}
\description{
Perform procrutes test quickly and tidy up the data to
data frame.
}
\examples{
\dontrun{
library(vegan)
data("varespec")
data("varechem")
procrutes_test(varespec, varechem)
procrutes_test(varespec, varechem, procrutes.fun = "procuste.randtest")
procrutes_test(varespec, varechem,
            spec.select = list(spec01 = 1:6, spec02 = 7:12))
procrutes_test(varespec, varechem, spec.pre.fun = "mono_mds",
            spec.select = list(spec01 = 1:6, spec02 = 7:12),
            env.select = list(env01 = 1:4, env02 = 5:14))
}
}
\seealso{
\code{\link[vegan]{protest}}, \code{\link[ade4]{procuste.rtest}},
\code{\link[ade4]{procuste.randtest}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
