% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcor.R
\name{quickcor}
\alias{quickcor}
\alias{print.quickcor}
\title{Plot Correlation Matrix Quickly}
\usage{
quickcor(
  x,
  y = NULL,
  mapping = NULL,
  grid.colour = "grey50",
  grid.size = 0.25,
  fixed.xy = TRUE,
  drop = TRUE,
  legend.position = "auto",
  ...
)

\method{print}{quickcor}(
  x,
  colours = getOption("ggcor.fill.pal"),
  style = getOption("ggcor.plot.style", "corrplot"),
  title = "corr",
  breaks = c(-1, -0.5, 0, 0.5, 1),
  labels = c(-1, -0.5, 0, 0.5, 1),
  limits = c(-1, 1),
  nbin = 40,
  ...
)
}
\arguments{
\item{x, y}{matrix, data frame or quickcor object in \code{print()}.}

\item{mapping}{NULL (default) or a list of aesthetic mappings to use for plot.}

\item{grid.colour}{colour of grid lines.}

\item{grid.size}{size of grid lines.}

\item{fixed.xy}{if TRUE (default), the coordinates will with fixed aspect ratio.}

\item{drop}{logical value, if TRUE (default) will drop the unused factor levels.}

\item{legend.position}{position of legend.}

\item{...}{ignore.}

\item{colours}{colour palette for filling.}

\item{style}{style of plot, one of "corrplot" (default) or "ggplot".}

\item{title}{guide title.}

\item{breaks}{breaks of guide_colourbar.}

\item{labels}{labels of guide_colourbar.}

\item{limits}{limits of guide_colourbar.}

\item{nbin}{a numeric specifying the number of bins for drawing the guide_colourbar.}
}
\description{
quickcor is convenient wrapper for creating a number of different types
of correlation matrix plots because of adding some extra settings by default.
}
\examples{
require(ggplot2, quietly = TRUE)

# Initialize the plot
quickcor(mtcars)

# layer of tile
quickcor(mtcars) + geom_colour()

# layer of circle and trim the lower triangle
quickcor(mtcars, type = "upper") + geom_circle2()

# layer of ellipse and not show diagonal
quickcor(mtcars, type = "lower", show.diag = FALSE) + geom_ellipse2()

# layer of square and reorder correlation matrix by cluster
quickcor(mtcars, cluster = TRUE) + geom_square()

# layer of confidence box
quickcor(mtcars, cor.test = TRUE) + geom_confbox()

# different layer of upper/lower triangle
quickcor(mtcars, cor.test = TRUE) +
  geom_square(data = get_data(type = "lower", show.diag = FALSE)) +
  geom_mark(data = get_data(type = "upper", show.diag = FALSE)) +
  geom_abline(slope = -1, intercept = 12)

## print quickcor object
p <- quickcor(mtcars) + geom_colour()
col <- c("blue", "white", "red")
print(p, colours = col)
print(p, colours = col, title = "Pearson's r")
print(p, style = "ggplot")
}
\seealso{
\code{\link[ggcor]{fortify_cor}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
