% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcor.R
\name{quickcor}
\alias{quickcor}
\alias{print.quickcor}
\alias{plot.quickcor}
\alias{is_quickcor}
\title{Plot Correlation Matrix Quickly}
\usage{
quickcor(
  x,
  y = NULL,
  mapping = NULL,
  grid.colour = "grey50",
  grid.size = 0.25,
  circular = FALSE,
  open = 90,
  inner = 1,
  outer = 0.3,
  fixed.xy = TRUE,
  axis.x.position = "auto",
  axis.y.position = "auto",
  axis.label.drop = TRUE,
  legend.position = "auto",
  bcols,
  paxis,
  ...
)

\method{print}{quickcor}(x, ...)

\method{print}{quickcor}(x, ...)

is_quickcor(x)
}
\arguments{
\item{x, y}{matrix, data frame or quickcor object in \code{print()}.}

\item{mapping}{NULL (default) or a list of aesthetic mappings to use for plot.}

\item{grid.colour}{colour of grid lines.}

\item{grid.size}{size of grid lines.}

\item{circular}{logical, if TRUE will draw in polar coordinates.}

\item{open}{angle of opening (in degree).}

\item{inner, outer}{the ratio of inner circle and outer margin.}

\item{fixed.xy}{if TRUE (default), the coordinates will with fixed aspect ratio.}

\item{axis.x.position, axis.y.position}{the position of the axis. 'auto' (default)
is set according to the plot type, 'bottom' or 'top' for x axes, 'left' or 'right'
for y axes.}

\item{axis.label.drop}{logical value (default is TRUE). When type of plot is 'upper'
or 'lower' and 'show.diag' is FALSE, do you need to remove the blank coordinate
label.}

\item{legend.position}{position of legend.}

\item{bcols}{ignore.}

\item{paxis}{ignore.}

\item{...}{extra params for \code{\link[ggcor]{fortify_cor}}.}
}
\description{
quickcor is convenient wrapper for creating a number of different types
of correlation matrix plots because of adding some extra settings by default.
}
\examples{
require(ggplot2, quietly = TRUE)

# Initialize the plot
quickcor(mtcars)

# layer of tile
quickcor(mtcars) + geom_colour()

# layer of circle and trim the lower triangle
quickcor(mtcars, type = "upper") + geom_circle2()

# layer of ellipse and not show diagonal
quickcor(mtcars, type = "lower", show.diag = FALSE) + geom_ellipse2()

# layer of square and reorder correlation matrix by cluster
quickcor(mtcars, cluster = TRUE) + geom_square()

# layer of confidence box
quickcor(mtcars, cor.test = TRUE) + geom_confbox()

# different layer of upper/lower triangle
quickcor(mtcars, cor.test = TRUE) +
  geom_square(data = get_data(type = "lower", show.diag = FALSE)) +
  geom_mark(data = get_data(type = "upper", show.diag = FALSE)) +
  geom_abline(slope = -1, intercept = 12)

}
\seealso{
\code{\link{fortify_cor}}.
}
\author{
Houyun Huang, Lei Zhou, Jian Chen, Taiyun Wei
}
