% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-purrr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{mutate}
\alias{mutate_if}
\alias{mutate_all}
\alias{mutate_each}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_if}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{mutate_each}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

