% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-graph.R, R/dplyr-purrr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.igraph}
\alias{plot.igraph}
\alias{as_tbl_graph}
\alias{\%>\%}
\alias{mutate}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{mutate}}}

  \item{igraph}{\code{\link[igraph]{as.igraph}}, \code{\link[igraph]{plot.igraph}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidygraph}{\code{\link[tidygraph]{as_tbl_graph}}}
}}

