% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-axis.R
\name{remove_axis}
\alias{remove_axis}
\title{Remove Axis Elements.}
\usage{
remove_axis(index = c("all", "x", "y"))
}
\arguments{
\item{index}{the index ("all", "x", "y") of axis will be removed.}
}
\value{
The theme.
}
\description{
A simple wrapper of the `theme()` function to quickly 
    remove axis elements.
}
