% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-gradient2n.R
\name{scale_colour_gradient2n}
\alias{scale_colour_gradient2n}
\alias{scale_color_gradient2n}
\alias{scale_fill_gradient2n}
\alias{scale_fill_steps2n}
\alias{scale_colour_steps2n}
\alias{scale_color_steps2n}
\title{Gradient n colour scales}
\usage{
scale_colour_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "colour")

scale_color_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "colour")

scale_fill_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "fill", colors)

scale_fill_steps2n(..., colours, midpoint = 0, values = NULL,
  space = "Lab", limits = c(-1, 1), na.value = "grey50",
  guide = "coloursteps", aesthetics = "fill", colors)

scale_colour_steps2n(..., colours, midpoint = 0, values = NULL,
  space = "Lab", limits = c(-1, 1), na.value = "grey50",
  guide = "coloursteps", aesthetics = "colour", colors)

scale_color_steps2n(..., colours, midpoint = 0, values = NULL,
  space = "Lab", limits = c(-1, 1), na.value = "grey50",
  guide = "coloursteps", aesthetics = "colour", colors)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}
\describe{
  \item{palette}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take (e.g., \code{\link[scales:hue_pal]{scales::hue_pal()}}).}
  \item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
  \item{limits}{A character vector that defines possible values of the scale
and their order.}
  \item{drop}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
  \item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
  \item{na.value}{If \code{na.translate = TRUE}, what value aesthetic
value should missing be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
  \item{scale_name}{The name of the scale that should be used for error messages
associated with this scale.}
  \item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
  \item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
  \item{expand}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
  \item{guide}{A function used to create a guide or its name. See
\code{\link[=guides]{guides()}} for more information.}
  \item{position}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}
  \item{super}{The super class to use for the constructed scale}
}}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use `"colourbar"` for continuous
colour bar, or `"legend"` for discrete colour legend.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{...}{Arguments passed on to \code{ggplot2::continuous_scale}
\describe{
  \item{scale_name}{The name of the scale that should be used for error messages
associated with this scale.}
  \item{palette}{A palette function that when called with a numeric vector with
values between 0 and 1 returns the corresponding output values
(e.g., \code{\link[scales:area_pal]{scales::area_pal()}}).}
  \item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
  \item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:trans_new]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:extended_breaks]{scales::extended_breaks()}})
}}
  \item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks.
}}
  \item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
  \item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[=coord_cartesian]{coord_cartesian()}}).
}}
  \item{rescaler}{A function used to scale the input values to the
range [0, 1]. This is always \code{\link[scales:rescale]{scales::rescale()}}, except for
diverging and n colour gradients (i.e., \code{\link[=scale_colour_gradient2]{scale_colour_gradient2()}},
\code{\link[=scale_colour_gradientn]{scale_colour_gradientn()}}). The \code{rescaler} is ignored by position
scales, which ways use \code{\link[scales:rescale]{scales::rescale()}}.}
  \item{oob}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds).
\item The default (\code{\link[scales:censor]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:squish]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:squish_infinite]{scales::squish_infinite()}} for squishing infitite values into range.
}}
  \item{expand}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
  \item{trans}{For continuous scales, the name of a transformation object
or the object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \code{<name>_trans} (e.g.,
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}). You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}
  \item{position}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}
  \item{super}{The super class to use for the constructed scale}
}}
}
\description{
`scale_*_gradient2n` creates a n colour gradient (low-high), mixed some
features of `scale_*_gradient2` and `scale_*_gradientn`.
}
\details{
Default colours are `c("#67001F", "#B2182B", "#D6604D", "#F4A582","#FDDBC7",
"#FFFFFF", "#D1E5F0", "#92C5DE","#4393C3", "#2166AC", "#053061")` , borrowed
from \pkg{corrplot}.
}
\examples{
df <- data.frame(x = rep(1:10, 10),
                 y = rep(1:10, each = 10),
                 z = runif(100, -1, 1))
library(ggplot2)
ggplot(df, aes(x, y, fill = z)) +
       geom_tile( ) +
       scale_gradient2n( )

ggplot(df, aes(x, y, colour = z)) +
       geom_point(size = 4) +
       scale_colour_gradient2n( )
}
\seealso{
[scales::seq_gradient_pal()] for details on underlying
  palette
}
