% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-gradient2n.R
\name{scale_colour_gradient2n}
\alias{scale_colour_gradient2n}
\alias{scale_color_gradient2n}
\alias{scale_fill_gradient2n}
\title{Gradient n colour scales}
\usage{
scale_colour_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "colour", colors = NULL)

scale_color_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "colour", colors = NULL)

scale_fill_gradient2n(..., colours, midpoint = 0, limits = NULL,
  space = "Lab", values = NULL, na.value = "grey50",
  guide = "colourbar", aesthetics = "fill", colors = NULL)
}
\arguments{
\item{...}{Arguments passed on to \code{continuous_scale}
\describe{
  \item{scale_name}{The name of the scale}
  \item{palette}{A palette function that when called with a numeric vector with
values between 0 and 1 returns the corresponding values in the range the
scale maps to.}
  \item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
  \item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output
}}
  \item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks.
}}
  \item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
  \item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits
}}
  \item{rescaler}{Used by diverging and n colour gradients
(i.e. \code{\link[=scale_colour_gradient2]{scale_colour_gradient2()}}, \code{\link[=scale_colour_gradientn]{scale_colour_gradientn()}}).
A function used to scale the input values to the range [0, 1].}
  \item{oob}{Function that handles limits outside of the scale limits
(out of bounds). The default replaces out of bounds values with \code{NA}.}
  \item{trans}{Either the name of a transformation object, or the
object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \code{name_trans}, e.g.
\code{\link[scales:boxcox_trans]{scales::boxcox_trans()}}. You can create your own
transformation with \code{\link[scales:trans_new]{scales::trans_new()}}.}
  \item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
  \item{super}{The super class to use for the constructed scale}
  \item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
}}

\item{colours}{Vector of colours to use for n-colour gradient.}

\item{midpoint}{The midpoint (in data value) of the diverging scale.
Defaults to 0.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link[=rescale]{rescale()}} for a convenience function
to map an arbitrary range to between 0 and 1.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{colors}{Vector of colours to use for n-colour gradient.}
}
\description{
`scale_*_gradient2n` creates a n colour gradient (low-high), mixed some
features of `scale_*_gradient2` and `scale_*_gradientn`.
}
\examples{
df <- data.frame(x = rep(1:10, 10),
                 y = rep(1:10, each = 10),
                 z = runif(100, -1, 1))
library(ggplot2)
ggplot(df, aes(x, y, fill = z)) +
       geom_tile( ) +
       scale_fill_gradient2n( )

ggplot(df, aes(x, y, colour = z)) +
       geom_point(size = 4) +
       scale_colour_gradient2n( )
}
\seealso{
[scales::seq_gradient_pal()] for details on underlying
  palette
}
